Module diffgeom::coordinates

Module containing basic types representing coordinate systems.

Structs

Point
Struct representing a point on the manifold. The information about the coordinate system is saved in the type parameter, so that only operations on objects belonging to the same coordinate system will be allowed.

Traits

ConversionTo
Trait used for conversions between different coordinate systems. Implementing ConversionTo<T> for a CoordinateSystem will allow objects in that system to be converted to the system T (note that T also has to be a CoordinateSystem).

CoordinateSystem
CoordinateSystem marks a struct (usually a unit struct) as representing a coordinate system.