Struct diffgeom::coordinates::Point

pub struct Point<T: CoordinateSystem> {
    // some fields omitted
}

Struct representing a point on the manifold. The information about the coordinate system is saved in the type parameter, so that only operations on objects belonging to the same coordinate system will be allowed.

Methods

impl<T> Point<T> where T: CoordinateSystem

fn new(coords: GenericArray<f64, T::Dimension>) -> Point<T>

Creates a new point with coordinates described by the array

fn from_slice(coords: &[f64]) -> Point<T>

Creates a new point with coordinates passed in the slice