Trait diffgeom::tensors::InnerProduct

pub trait InnerProduct<Rhs, Ul: Unsigned, Uh: Unsigned> {
    type Output;
    fn inner_product(self, rhs: Rhs) -> Self::Output;
}

Trait representing the inner product of two tensors.

The inner product is just a multiplication followed by a contraction. The contraction is defined by type parameters Ul and Uh. Ul has to be less than Uh and the indices at those positions must be of opposite types (checked at compile time)

Associated Types

type Output

Required Methods

fn inner_product(self, rhs: Rhs) -> Self::Output