Trait diffgeom::coordinates::CoordinateSystem

pub trait CoordinateSystem: Sized {
    type Dimension: Unsigned + ArrayLength<f64> + ArrayLength<usize>;
    fn small(_: &Point<Self>) -> f64 { ... }
    fn dimension() -> usize { ... }
}

CoordinateSystem marks a struct (usually a unit struct) as representing a coordinate system.

Associated Types

type Dimension: Unsigned + ArrayLength<f64> + ArrayLength<usize>

An associated type representing the dimension of the coordinate system

Provided Methods

fn small(_: &Point<Self>) -> f64

Function returning a small value for purposes of numerical differentiation. What is considered a small value may depend on the point, hence the parameter. Returns just 0.01 by default.

fn dimension() -> usize

Function returning the dimension